services/ml/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.0
pillow>=11.0
numpy>=1.26.0
